requirements/base.txt
faust
simple-settings
Pillow
